+++ spec/update_assigned_milestone_spec.rb
@@ -72,8 +72,12 @@ describe Fastlane::Actions::UpdateAssignedMilestoneAction do
   context 'when providing a source milestone' do
     it 'updates the milestone of all matching and still-opened PRs' do
       allow(client).to receive(:search_issues)
-        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:open))
-        .and_return({ items: [101, 103].map { |n| mock_pr(n) } })
+        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:issue is:open))
+        .and_return({ items: [101].map { |n| mock_pr(n) } })
+
+      allow(client).to receive(:search_issues)
+        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:pull-request is:open))
+        .and_return({ items: [103].map { |n| mock_pr(n) } })
 
       expect(client).to receive(:update_issue).with(test_repo, 101, { milestone: 123 })
       expect(client).to receive(:update_issue).with(test_repo, 103, { milestone: 123 })
@@ -91,8 +95,13 @@ describe Fastlane::Actions::UpdateAssignedMilestoneAction do
     it 'adds a PR comment if one is provided' do
       comment = 'Updated milestone from `12.2` to `12.3`'
       allow(client).to receive(:search_issues)
-        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:open))
-        .and_return({ items: [101, 103].map { |n| mock_pr(n) } })
+        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:issue is:open))
+        .and_return({ items: [101].map { |n| mock_pr(n) } })
+
+      allow(client).to receive(:search_issues)
+        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:pull-request is:open))
+        .and_return({ items: [103].map { |n| mock_pr(n) } })
+
       allow(client).to receive(:issue_comments).and_return([])
 
       expect(client).to receive(:update_issue).with(test_repo, 101, { milestone: 123 })
@@ -113,8 +122,12 @@ describe Fastlane::Actions::UpdateAssignedMilestoneAction do
 
     it 'does not add a PR comment if comment is empty' do
       allow(client).to receive(:search_issues)
-        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:open))
-        .and_return({ items: [101, 103].map { |n| mock_pr(n) } })
+        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:issue is:open))
+        .and_return({ items: [101].map { |n| mock_pr(n) } })
+
+      allow(client).to receive(:search_issues)
+        .with(%(repo:#{test_repo} milestone:"#{mock_milestone(12.2)[:title]}" is:pull-request is:open))
+        .and_return({ items: [103].map { |n| mock_pr(n) } })
 
       expect(client).to receive(:update_issue).with(test_repo, 101, { milestone: 123 })
       expect(client).to receive(:update_issue).with(test_repo, 103, { milestone: 123 })